Fixed bug in Win32 version of getmilliseconds: Visual C 6.0 doesn't understand "unsigned long long" so we use LONGLONG defined in the Win32 SDK headers instead.
git-svn-id: https://frontierkernel.svn.sourceforge.net/svnroot/frontierkernel/Frontier/trunk@393 4e289971-d10f-0410-980f-25908192c0f7

+++ Common/source/timedate.c
@@ -1554,7 +1554,7 @@ long getcurrenttimezonebias(void) {
 
 extern long getmilliseconds(void) {
 	LARGE_INTEGER counter, freq;
-	unsigned long long freqLong;
+	LONGLONG freqLong;
 	
 	QueryPerformanceCounter (&counter);  /* counter hits since boot-up */
